requirements.txt
numpy
PyOpenGL
PyOpenGL-accelerate
pygame

